@@ -1,30 +1,28 @@
 {
-    "name": "vue-vite-plugin",
-    "version": "0.0.1",
-    "description": "",
-    "main": "index.js",
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "start": "vite"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/btea/vue-vite-plugin.git"
-    },
-    "author": "btea",
-    "license": "MIT",
-    "bugs": {
-        "url": "https://github.com/btea/vue-vite-plugin/issues"
-    },
-    "homepage": "https://github.com/btea/vue-vite-plugin#readme",
-    "dependencies": {
-        "@vue/composition-api": "^1.1.3",
-        "vue": "^2.6.14"
-    },
-    "devDependencies": {
-        "vite": "^2.5.0",
-        "vite-plugin-vue2": "^1.8.1",
-        "vue-template-compiler": "^2.6.14",
-        "vue2-script-setup-transform": "^0.3.5"
-    }
+	"name": "vue-vite-plugin",
+	"version": "0.0.1",
+	"description": "",
+	"main": "index.js",
+	"scripts": {
+		"test": "echo \"Error: no test specified\" && exit 1",
+		"start": "vite"
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/btea/vue-vite-plugin.git"
+	},
+	"author": "btea",
+	"license": "MIT",
+	"bugs": {
+		"url": "https://github.com/btea/vue-vite-plugin/issues"
+	},
+	"homepage": "https://github.com/btea/vue-vite-plugin#readme",
+	"dependencies": {
+		"@vitejs/plugin-vue": "^1.10.0",
+		"vue": "^3.2.22"
+	},
+	"devDependencies": {
+		"unplugin-vue-components": "^0.17.2",
+		"vite": "^2.6.14"
+	}
 }
