@@ -1,9 +1,4 @@
-import Vue from 'vue';
-import VueComposition from '@vue/composition-api';
+import { createApp } from "vue";
+import App from './App.vue'
 
-Vue.use(VueComposition);
-import App from './App.vue';
-console.log(App);
-new Vue({
-	render: h => h(App),
-}).$mount('#app');
+createApp(App).mount('#app')
